Removed a compilation warning.

+++ src/s_font.cpp
@@ -132,7 +132,8 @@ float s_font::getStrLen(const string& str) const
     float s=0;
     for (unsigned int i=0;i<str.length();++i)
     { 
-		s+=theSize[str[i]].sizeX+theSize[str[i]].leftSpacing+theSize[str[i]].rightSpacing+SPACING;
+		s += theSize[(unsigned int)str[i]].sizeX + theSize[(unsigned int)str[i]].leftSpacing +
+			theSize[(unsigned int)str[i]].rightSpacing + SPACING;
     }
     return s*ratio;
 }
@@ -142,7 +143,7 @@ float s_font::getStrHeight(const string& str) const
     float s=0;
     for (unsigned int i=0;i<str.length();++i)
     {
-		if (s<theSize[str[i]].sizeY) s=theSize[str[i]].sizeY;
+		if (s<theSize[(unsigned int)str[i]].sizeY) s=theSize[(unsigned int)str[i]].sizeY;
     }
     return s*ratio;
 }
